some design changes and fixes

+++ Entities/Views/JobOverview.cs
@@ -11,7 +11,7 @@ namespace HlidacStatu.Entities.Views
         
         public int Year { get; set; }
         
-        public int JobPk { get; set; }
+        public long JobPk { get; set; }
         public string JobGrouped { get; set; }
         public decimal? SalaryMd { get; set; }
         public decimal? SalaryMdVat { get; set; }
